JNI: cast all int to jint in com_wolfssl_WolfSSLContext.c

+++ native/com_wolfssl_WolfSSLContext.c
@@ -176,7 +176,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setTmpDH
     (void)jcl;
 
     if (jenv == NULL || p == NULL || g == NULL) {
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
     }
 
     if (ctx == NULL) {
@@ -184,12 +184,12 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setTmpDH
         if ((*jenv)->ExceptionOccurred(jenv)) {
             (*jenv)->ExceptionDescribe(jenv);
             (*jenv)->ExceptionClear(jenv);
-            return SSL_FAILURE;
+            return (jint)SSL_FAILURE;
         }
         (*jenv)->ThrowNew(jenv, excClass,
                 "Input WolfSSLContext object was null in "
                 "setTmpDH");
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     pBuf = (unsigned char*)XMALLOC((int)pSz, NULL, DYNAMIC_TYPE_TMP_BUFFER);
@@ -233,7 +233,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setTmpDH
     XFREE(pBuf, NULL, DYNAMIC_TYPE_TMP_BUFFER);
     XFREE(gBuf, NULL, DYNAMIC_TYPE_TMP_BUFFER);
 
-    return ret;
+    return (jint)ret;
 #else
     (void)jenv;
     (void)jcl;
@@ -242,7 +242,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setTmpDH
     (void)pSz;
     (void)g;
     (void)gSz;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif /* !NO_DH */
 }
 
@@ -257,7 +257,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setTmpDHFile
     (void)jcl;
 
     if (file == NULL) {
-        return SSL_BAD_FILE;
+        return (jint)SSL_BAD_FILE;
     }
 
     if (ctx == NULL) {
@@ -265,12 +265,12 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setTmpDHFile
         if ((*jenv)->ExceptionOccurred(jenv)) {
             (*jenv)->ExceptionDescribe(jenv);
             (*jenv)->ExceptionClear(jenv);
-            return SSL_FAILURE;
+            return (jint)SSL_FAILURE;
         }
         (*jenv)->ThrowNew(jenv, excClass,
                 "Input WolfSSLContext object was null in "
                 "setTmpDHFile");
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     fname = (*jenv)->GetStringUTFChars(jenv, file, 0);
@@ -279,14 +279,14 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setTmpDHFile
 
     (*jenv)->ReleaseStringUTFChars(jenv, file, fname);
 
-    return ret;
+    return (jint)ret;
 #else
     (void)jenv;
     (void)jcl;
     (void)ctxPtr;
     (void)file;
     (void)format;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif /* !NO_DH && !NO_FILESYSTEM */
 }
 
@@ -301,7 +301,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_useCertificateFile
     (void)jcl;
 
     if (jenv == NULL) {
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     if (file == NULL) {
@@ -314,7 +314,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_useCertificateFile
         (*jenv)->ThrowNew(jenv, excClass,
                 "Input certificate file is NULL");
 
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     certFile = (*jenv)->GetStringUTFChars(jenv, file, 0);
@@ -330,7 +330,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_useCertificateFile
     (void)ctxPtr;
     (void)file;
     (void)format;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -345,7 +345,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_usePrivateKeyFile
     (void)jcl;
 
     if (jenv == NULL) {
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     if (file == NULL) {
@@ -358,7 +358,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_usePrivateKeyFile
         (*jenv)->ThrowNew(jenv, excClass,
                 "Input private key file is NULL");
 
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     keyFile = (*jenv)->GetStringUTFChars(jenv, file, 0);
@@ -374,7 +374,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_usePrivateKeyFile
     (void)ctxPtr;
     (void)file;
     (void)format;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -390,7 +390,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_loadVerifyLocations
     (void)jcl;
 
     if (jenv == NULL) {
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     if (file == NULL && path == NULL) {
@@ -403,7 +403,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_loadVerifyLocations
         (*jenv)->ThrowNew(jenv, excClass,
                 "Input file and path are both NULL");
 
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     if (file) {
@@ -432,7 +432,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_loadVerifyLocations
     (void)ctxPtr;
     (void)file;
     (void)path;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -447,7 +447,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_useCertificateChainFile
     (void)jcl;
 
     if (jenv == NULL) {
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     /* throw exception if no input file */
@@ -461,7 +461,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_useCertificateChainFile
         (*jenv)->ThrowNew(jenv, excClass,
                 "Input certificate chain file is NULL");
 
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     chainFile = (*jenv)->GetStringUTFChars(jenv, file, 0);
@@ -476,7 +476,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_useCertificateChainFile
     (void)jcl;
     (void)ctxPtr;
     (void)file;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -638,7 +638,7 @@ JNIEXPORT jlong JNICALL Java_com_wolfssl_WolfSSLContext_setOptions
         return 0;
     }
 
-    return wolfSSL_CTX_set_options(ctx, (long)op);
+    return (jlong)wolfSSL_CTX_set_options(ctx, (long)op);
 }
 
 JNIEXPORT jlong JNICALL Java_com_wolfssl_WolfSSLContext_getOptions
@@ -651,7 +651,7 @@ JNIEXPORT jlong JNICALL Java_com_wolfssl_WolfSSLContext_getOptions
         return 0;
     }
 
-    return wolfSSL_CTX_get_options(ctx);
+    return (jlong)wolfSSL_CTX_get_options(ctx);
 }
 
 JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_memsaveCertCache
@@ -667,7 +667,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_memsaveCertCache
     (void)jcl;
 
     if (jenv == NULL || ctx == NULL || mem == NULL || sz <= 0) {
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
     }
 
     /* find exception class */
@@ -675,12 +675,12 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_memsaveCertCache
     if ((*jenv)->ExceptionOccurred(jenv)) {
         (*jenv)->ExceptionDescribe(jenv);
         (*jenv)->ExceptionClear(jenv);
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     memBuf = (unsigned char*)XMALLOC((int)sz, NULL, DYNAMIC_TYPE_TMP_BUFFER);
     if (memBuf == NULL) {
-        return MEMORY_E;
+        return (jint)MEMORY_E;
     }
     XMEMSET(memBuf, 0, (int)sz);
 
@@ -697,7 +697,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_memsaveCertCache
         (*jenv)->ThrowNew(jenv, excClass,
                 "Failed to set array region in native memsaveCertCache");
 
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     /* set jbyteArray for return */
@@ -712,13 +712,13 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_memsaveCertCache
             (*jenv)->ThrowNew(jenv, excClass,
                     "Failed to set byte region in native memsaveCertCache");
 
-            return SSL_FAILURE;
+            return (jint)SSL_FAILURE;
         }
     }
 
     XFREE(memBuf, NULL, DYNAMIC_TYPE_TMP_BUFFER);
 
-    return ret;
+    return (jint)ret;
 #else
     (void)jenv;
     (void)jcl;
@@ -726,7 +726,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_memsaveCertCache
     (void)mem;
     (void)sz;
     (void)used;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -761,7 +761,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_memrestoreCertCache
     (void)ctxPtr;
     (void)mem;
     (void)sz;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -772,9 +772,9 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_getCertCacheMemsize
     (void)jcl;
 #ifdef PERSIST_CERT_CACHE
     /* wolfSSL checks for null pointer */
-    return wolfSSL_CTX_get_cert_cache_memsize((WOLFSSL_CTX*)(uintptr_t)ctx);
+    return (jint)wolfSSL_CTX_get_cert_cache_memsize((WOLFSSL_CTX*)(uintptr_t)ctx);
 #else
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -784,7 +784,8 @@ JNIEXPORT jlong JNICALL Java_com_wolfssl_WolfSSLContext_setCacheSize
     (void)jenv;
     (void)jcl;
 
-    return wolfSSL_CTX_sess_set_cache_size((WOLFSSL_CTX*)(uintptr_t)ctx, (long)sz);
+    return (jlong)wolfSSL_CTX_sess_set_cache_size(
+                    (WOLFSSL_CTX*)(uintptr_t)ctx, (long)sz);
 }
 
 
@@ -794,7 +795,7 @@ JNIEXPORT jlong JNICALL Java_com_wolfssl_WolfSSLContext_getCacheSize
     (void)jenv;
     (void)jcl;
 
-    return wolfSSL_CTX_sess_get_cache_size((WOLFSSL_CTX*)(uintptr_t)ctx);
+    return (jlong)wolfSSL_CTX_sess_get_cache_size((WOLFSSL_CTX*)(uintptr_t)ctx);
 }
 
 JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setCipherList
@@ -807,7 +808,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setCipherList
     (void)jcl;
 
     if (jenv == NULL) {
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     if (list == NULL) {
@@ -821,7 +822,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setCipherList
         (*jenv)->ThrowNew(jenv, excClass,
                 "Input cipher list is NULL");
 
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     cipherList = (*jenv)->GetStringUTFChars(jenv, list, 0);
@@ -976,7 +977,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setGroupMessages
         return (jint)BAD_FUNC_ARG;
     }
 
-    return wolfSSL_CTX_set_group_messages((WOLFSSL_CTX*)(uintptr_t)ctx);
+    return (jint)wolfSSL_CTX_set_group_messages((WOLFSSL_CTX*)(uintptr_t)ctx);
 }
 
 JNIEXPORT void JNICALL Java_com_wolfssl_WolfSSLContext_setIORecv(JNIEnv* jenv,
@@ -1660,11 +1661,11 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_enableCRL
     (void)jcl;
 #ifdef HAVE_CRL
     if (!ctx)
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
 
-    return wolfSSL_CTX_EnableCRL((WOLFSSL_CTX*)(uintptr_t)ctx, options);
+    return (jint)wolfSSL_CTX_EnableCRL((WOLFSSL_CTX*)(uintptr_t)ctx, options);
 #else
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -1675,11 +1676,11 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_disableCRL
     (void)jcl;
 #ifdef HAVE_CRL
     if (!ctx)
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
 
-    return wolfSSL_CTX_DisableCRL((WOLFSSL_CTX*)(uintptr_t)ctx);
+    return (jint)wolfSSL_CTX_DisableCRL((WOLFSSL_CTX*)(uintptr_t)ctx);
 #else
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -1694,7 +1695,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_loadCRL
     (void)jcl;
 
     if (jenv == NULL || ctx == NULL || path == NULL) {
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
     }
 
     crlPath = (*jenv)->GetStringUTFChars(jenv, path, 0);
@@ -1703,7 +1704,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_loadCRL
 
     (*jenv)->ReleaseStringUTFChars(jenv, path, crlPath);
 
-    return ret;
+    return (jint)ret;
 #else
     (void)jenv;
     (void)jcl;
@@ -1711,7 +1712,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_loadCRL
     (void)path;
     (void)type;
     (void)monitor;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -1725,7 +1726,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setCRLCb
     (void)jcl;
 
     if (jenv == NULL || ctx == NULL) {
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
     }
 
     /* release global CRL callback object if set */
@@ -1752,13 +1753,13 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setCRLCb
         ret = wolfSSL_CTX_SetCRL_Cb(ctx, NativeCtxMissingCRLCallback);
     }
 
-    return ret;
+    return (jint)ret;
 #else
     (void)jenv;
     (void)jcl;
     (void)ctxPtr;
     (void)cb;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -1855,9 +1856,10 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_enableOCSP
     (void)jcl;
 #ifdef HAVE_OCSP
     /* wolfSSL checks for null pointer */
-    return wolfSSL_CTX_EnableOCSP((WOLFSSL_CTX*)(uintptr_t)ctx, (long)options);
+    return (jint)wolfSSL_CTX_EnableOCSP((WOLFSSL_CTX*)(uintptr_t)ctx,
+                                        (long)options);
 #else
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -1868,9 +1870,9 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_disableOCSP
     (void)jcl;
 #ifdef HAVE_OCSP
     /* wolfSSL checks for null pointer */
-    return wolfSSL_CTX_DisableOCSP((WOLFSSL_CTX*)(uintptr_t)ctx);
+    return (jint)wolfSSL_CTX_DisableOCSP((WOLFSSL_CTX*)(uintptr_t)ctx);
 #else
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -1886,7 +1888,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setOCSPOverrideUrl
 
     /* wolfSSL checks ctx for NULL */
     if (jenv == NULL)
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
 
     if (urlString == NULL)
     {
@@ -1900,7 +1902,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setOCSPOverrideUrl
         (*jenv)->ThrowNew(jenv, excClass,
                 "Input URL is NULL in setOCSPOverrideUrl()");
 
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     url = (*jenv)->GetStringUTFChars(jenv, urlString, 0);
@@ -1915,7 +1917,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setOCSPOverrideUrl
     (void)jcl;
     (void)ctxPtr;
     (void)urlString;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -5424,7 +5426,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_usePskIdentityHint
     (void)obj;
 
     if (jenv == NULL || ctx == NULL || hint == NULL) {
-        return SSL_FAILURE;
+        return (jint)SSL_FAILURE;
     }
 
     nativeHint = (*jenv)->GetStringUTFChars(jenv, hint, 0);
@@ -5439,7 +5441,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_usePskIdentityHint
     (void)obj;
     (void)ctxPtr;
     (void)hint;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -5453,7 +5455,7 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_useSecureRenegotiation
                     (WOLFSSL_CTX*)(uintptr_t)ctx);
 #else
     (void)ctx;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -5466,16 +5468,16 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setMinDhKeySz
 
     /* wolfSSL_CTX_SetMinDhKey_Sz() sanitizes ctx and keySzBits */
     if (jenv == NULL) {
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
     }
 
-    return wolfSSL_CTX_SetMinDhKey_Sz(ctx, (word16)keySzBits);
+    return (jint)wolfSSL_CTX_SetMinDhKey_Sz(ctx, (word16)keySzBits);
 #else
     (void)jenv;
     (void)obj;
     (void)ctxPtr;
     (void)keySzBits;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -5488,16 +5490,16 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setMinRsaKeySz
 
     /* wolfSSL_CTX_SetMinRsaKey_Sz() sanitizes ctx and keySzBits */
     if (jenv == NULL) {
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
     }
 
-    return wolfSSL_CTX_SetMinRsaKey_Sz(ctx, (short)keySzBits);
+    return (jint)wolfSSL_CTX_SetMinRsaKey_Sz(ctx, (short)keySzBits);
 #else
     (void)jenv;
     (void)obj;
     (void)ctxPtr;
     (void)keySzBits;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
 
@@ -5510,16 +5512,16 @@ JNIEXPORT jint JNICALL Java_com_wolfssl_WolfSSLContext_setMinEccKeySz
 
     /* wolfSSL_CTX_SetMinEccKey_Sz() sanitizes ctx and keySzBits */
     if (jenv == NULL) {
-        return BAD_FUNC_ARG;
+        return (jint)BAD_FUNC_ARG;
     }
 
-    return wolfSSL_CTX_SetMinEccKey_Sz(ctx, (short)keySzBits);
+    return (jint)wolfSSL_CTX_SetMinEccKey_Sz(ctx, (short)keySzBits);
 #else
     (void)jenv;
     (void)obj;
     (void)ctxPtr;
     (void)keySzBits;
-    return NOT_COMPILED_IN;
+    return (jint)NOT_COMPILED_IN;
 #endif
 }
